refactor(packaging): view/adapter
